+++ backend/functions/api/analytics.js
@@ -7,19 +7,14 @@ const {
   FOOD_CATEGORIES,
   RECIPIENT_SUB_TYPES,
   RETAIL_VALUES,
-} = require('./helpers')
-const express = require('express')
-const cors = require('cors')
+} = require('../../helpers')
 
-const analytics_routes = express()
-analytics_routes.use(cors({ origin: true }))
-analytics_routes.get('/', handleAnalytics)
-exports.analytics = analytics_routes
-
-async function handleAnalytics(request, response) {
+exports.analytics = async (request, response) => {
   return new Promise(async resolve => {
     console.log('running analytics')
-    const { date_range_start, date_range_end, breakdown } = request.query
+    const { date_range_start, date_range_end } = request.query
+    let { breakdown } = request.query
+    breakdown = decodeURIComponent(breakdown)
     console.log(date_range_start, date_range_end, breakdown)
 
     let rescues = []
