+++ backend/functions/index.js
@@ -1,0 +1,5 @@
+module.exports = {
+  ...require('./api'),
+  ...require('./backup_data_to_storage'),
+  ...require('./export_data_to_google_sheets'),
+}
