+++ backend/helpers/index.js
@@ -1,0 +1,1 @@
+module.exports = { ...require('./functions'), ...require('./constants') }
