+++ firebase.json
@@ -11,6 +11,6 @@
     ]
   },
   "functions": {
-    "source": "functions"
+    "source": "backend"
   }
 }
