--- functions/.eslintrc.json
@@ -1,6 +1,0 @@
-{
-  "rules": {
-    "no-var": "error",
-    "prefer-const": "error"
-  }
-}
