+++ package.json
@@ -8,6 +8,7 @@
     "@sentry/react": "^6.16.1",
     "@sentry/tracing": "^6.16.1",
     "@sharingexcess/designsystem": "^1.2.8",
+    "concurrently": "^7.0.0",
     "firebase": "^9.6.6",
     "google-map-react": "^2.1.10",
     "moment": "^2.29.1",
@@ -24,8 +25,10 @@
   },
   "scripts": {
     "start": "react-scripts start",
-    "dev": "cp environments/.env.dev .env.local && react-scripts start",
-    "prod": "cp environments/.env.prod .env.local && react-scripts start",
+    "set_env:dev": "cp backend/environments/.env.prod backend/.env && cp environments/.env.prod .env",
+    "set_env:prod": "cp backend/environments/.env.prod backend/.env && cp environments/.env.prod .env",
+    "dev": "npm run set_env:dev && concurrently \"npm run emulators\" \"npm start\"",
+    "prod": "npm run set_env:prod && concurrently \"npm run emulators\" \"npm start\"",
     "emulators": "firebase emulators:start --only functions",
     "build": "react-scripts build",
     "deploy": "node scripts/manual_prod_deploy.js",
