+++ src/components/DriverStats/DriverStats.js
@@ -1,30 +1,32 @@
 import { Spacer, Text } from '@sharingexcess/designsystem'
-import { ORG_SUBTYPES, STATUSES } from 'helpers'
-import { useAuth, useFirestore } from 'hooks'
-import React, { useCallback, useEffect, useMemo } from 'react'
+import { ORG_SUBTYPES } from 'helpers'
+import { useFirestore } from 'hooks'
+import React, { useMemo } from 'react'
 import { PoundsByMonthChart } from './PoundsByMonthChart'
 import { PoundsByOrgChart } from './PoundsByOrgChart'
 import { TotalPounds } from './TotalPounds'
 
 export function DriverStats() {
-  const { user } = useAuth()
-  const { loadAllData } = useFirestore()
-  const driver_stops = useFirestore(
-    'stops',
-    useCallback(
-      i =>
-        i.handler_id === user.id &&
-        // i.handler_id === 'jvC1BuuhYiXzMvbuog9b9YcUkDy1' && // (Use Jacob's ID for testing)
-        // i.handler_id === 'jiBBxAvncBSNWwjizdmsxM7EALz1' && // Sarah DiPasquale
-        // i.handler_id === '1yxUOZ53OOg0T3SJ2SqolPI3UK12' && // Alex Havertine
-        // [
-        //   'sUqgP36KAiPgDaU7V4z8N0GirIc2',
-        //   'K2eYju4PtaejTIiOdHQrERMoRnk2',
-        // ].includes(i.handler_id) && // Evan Ehlers (he has 2 accounts)
-        i.status === STATUSES.COMPLETED,
-      [user]
-    )
-  )
+  // const { user } = useAuth()
+  // const { loadAllData } = useFirestore()
+  // const driver_stops = useFirestore(
+  //   'stops',
+  //   useCallback(
+  //     i =>
+  //       i.handler_id === user.id &&
+  //       // i.handler_id === 'jvC1BuuhYiXzMvbuog9b9YcUkDy1' && // (Use Jacob's ID for testing)
+  //       // i.handler_id === 'jiBBxAvncBSNWwjizdmsxM7EALz1' && // Sarah DiPasquale
+  //       // i.handler_id === '1yxUOZ53OOg0T3SJ2SqolPI3UK12' && // Alex Havertine
+  //       // [
+  //       //   'sUqgP36KAiPgDaU7V4z8N0GirIc2',
+  //       //   'K2eYju4PtaejTIiOdHQrERMoRnk2',
+  //       // ].includes(i.handler_id) && // Evan Ehlers (he has 2 accounts)
+  //       i.status === STATUSES.COMPLETED,
+  //     [user]
+  //   )
+  // )
+  // TEMP FIX => ACTIVELY BEING REFACTORED IN ANOTHER PR
+  const driver_stops = [] // eslint-disable-line
   const organizations = useFirestore('organizations')
 
   const filteredByHolding = useMemo(
@@ -45,7 +47,7 @@ export function DriverStats() {
     [filteredByHolding]
   )
 
-  useEffect(() => loadAllData(), []) // eslint-disable-line
+  // useEffect(() => loadAllData(), []) // eslint-disable-line
 
   return (
     <main id="DriverStats">
