+++ src/helpers/firebase.js
@@ -2,7 +2,13 @@ import firebase from 'firebase/compat/app'
 import 'firebase/compat/auth'
 import 'firebase/compat/firestore'
 import { customAlphabet } from 'nanoid'
+import { FIREBASE_CONFIG } from './constants'
 const nanoid = customAlphabet('1234567890abcdefghijklmnopqrstuvwqyz', 12)
+
+// This function call connects us to Firebase and initializes all of our API access
+firebase.initializeApp(FIREBASE_CONFIG)
+
+export const firestore = firebase.firestore()
 
 export function generateId() {
   return nanoid()
@@ -15,15 +21,11 @@ export async function generateUniqueId(collection) {
 }
 
 export function getCollection(name) {
-  return firebase.firestore().collection(name)
+  return firestore.collection(name)
 }
 
 export async function isExistingId(id, collection) {
-  const snapshot = await firebase
-    .firestore()
-    .collection(collection)
-    .doc(id)
-    .get()
+  const snapshot = await firestore.collection(collection).doc(id).get()
   return snapshot.exists
 }
