+++ src/hooks/index.js
@@ -2,6 +2,7 @@ export * from './useIsMobile'
 export * from './useWidth'
 export * from './useAuth'
 export * from './useFirestore'
+export * from './useFirestoreListener'
 export * from './useScroll'
 export * from './useApp'
 export * from './useApi'
