--- ui-debug.log
@@ -1,1 +1,0 @@
-Web / API server started at localhost:4002
